30.03.2015.1

+++ inheritance2/src/by/gsu/epamlab/Commodity.java
@@ -3,13 +3,13 @@ package by.gsu.epamlab;
 public class Commodity {
 
 	private String name;
-	private double price;
+	private int price;
 	
 	public Commodity() {
 		super();
 	}
 
-	public Commodity(String name, double price) {
+	public Commodity(String name, int price) {
 		super();
 		this.name = name;
 		this.price = price;
@@ -23,11 +23,11 @@ public class Commodity {
 		this.name = name;
 	}
 
-	public double getPrice() {
+	public int getPrice() {
 		return price;
 	}
 
-	public void setPrice(double price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
